Update README.md
First, employees are defined, their addresses are stored in the address field. Addresses only
that are on that list can access the contract. Then three variables are defined:
totalReceivedMoney, which counts the total amount of money received, withdrawnAmounts, which records how much money each employee has withdrawn, and employees, which lists the employees who are authorized to withdraw money from the contract. The "Withdraw" function checks whether the address of the account requesting the transaction belongs to the list of addresses.
If it belongs, the function returns the value "true", otherwise it returns the value "false". In order for all employees to pay the fuel equally for the transaction, the "break" command is commented out. If the "break" command is not commented out, the first employee in the list will pay less fuel because the loop will be executed for a shorter time. This could lead to disparities in fuel payments between employees. The loop inside the Withdraw function checks whether the caller of the function (employee) belongs to the employees field, that is, whether he has the right to withdraw money.
If the caller of the function belongs to the employees field, then the amount of money it belongs to is determined.